Updated generics

+++ benchmark/dyn_array_push.cpp
@@ -7,7 +7,7 @@ int main() {
         vec.push_back(i);
     }
     long long sum = 0;
-    for (int i = 0; i < vec.size(); i++) {
+    for (size_t i = 0; i < vec.size(); i++) {
         sum += vec[i];
     }
     std::cout << "Sum: " << sum << std::endl;
